% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d}
\alias{d}
\title{This is data which is included in package, KangHanGyutools}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 142 rows and 6 columns.}
\usage{
d
}
\description{
This is data which is included in package, KangHanGyutools
}
\author{
HanGyu Kang \email{kangx478@umn.edu}
}
\keyword{datasets}
